+++ src/ops_data_store/config.py
@@ -1,6 +1,17 @@
 from importlib.metadata import version
+from os import environ
 
 from dotenv import load_dotenv
+
+
+def _get_env(env: str) -> str:
+    """
+    Get value from environment variables.
+
+    Util method to corrects for double wrapped (e.g. '"Foo"') values and/or values with leading or trailing whitespace.
+    Needed due to the way we override settings using an additional .env file for testing.
+    """
+    return environ.get(env).replace('"', "").strip()
 
 
 class Config:
@@ -12,9 +23,14 @@ class Config:
 
     def dump(self) -> dict:
         """Return application configuration as a dictionary."""
-        return {"VERSION": self.VERSION}
+        return {"VERSION": self.VERSION, "DB_DSN": self.DB_DSN}
 
     @property
     def VERSION(self) -> str:
         """Application version."""
         return version("ops-data-store")
+
+    @property
+    def DB_DSN(self) -> str:
+        """DB connection string."""
+        return _get_env("APP_ODS_DB_DSN")