+++ tests/ops_data_store_tests/cli/test_cli_config.py
@@ -1,4 +1,3 @@
-from importlib.metadata import version
 from io import StringIO
 from pprint import pprint
 
@@ -7,16 +6,16 @@ from typer.testing import CliRunner
 from ops_data_store.cli import app as cli
 
 
-class TestCli:
-    """Tests for `config` Typer CLI group."""
+class TestCliConfigShow:
+    """Tests for `config show` CLI command."""
 
-    def test_show(self, fx_cli_runner: CliRunner) -> None:
-        """The `show` command returns app config."""
+    def test_ok(self, fx_cli_runner: CliRunner, fx_test_package_version: str, fx_test_db_dsn: str) -> None:
+        """Returns app config."""
         result = fx_cli_runner.invoke(app=cli, args=["config", "show"])
 
-        # can't use `str(dict)` as output is formatted by pprint in CLI
+        # Can't simply use `str(dict)` as output is formatted by pprint in CLI so will contain new lines etc.
         output = StringIO()
-        pprint({"VERSION": version("ops-data-store")}, stream=output)
+        pprint({"VERSION": fx_test_package_version, "DB_DSN": fx_test_db_dsn}, stream=output)
         expected = output.getvalue()
 
         assert result.exit_code == 0