+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#14](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/14)
 * Documenting IT Python version
   [#4](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/4)
+* Documenting and including IT GDAL version
+  [#5](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/5)
 
 ## [0.1.0] - 2023-10-01